requirements.txt
discord.py>=2.0.0,<3.0.0
aiohttp>=3.8.0
python-dotenv>=0.21.0
asyncio
aiosqlite